test: remove sequencer which doesn't seem to work with esm modules and convince jest (aka. "use a bludgeon") to run sequential integration tests in the expected alphabetic file name order
